requirements.txt

torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
transformers==4.35.2
accelerate
safetensors
ftfy
regex
Pillow>=9.0.0
opencv-python>=4.8.0
scikit-learn>=1.0.0
tqdm>=4.64.0
matplotlib>=3.5.0
numpy>=1.21.0


# For CLIP
git+https://github.com/openai/CLIP.git

